updating post on creating post
