+++ src/components/posts/index.js
@@ -1,18 +1,6 @@
-import { useEffect, useState } from 'react';
 import Post from '../post';
-import { getPosts } from '../../service/apiClient';
 
-const Posts = () => {
-  const [posts, setPosts] = useState([]);
-
-  useEffect(() => {
-    getPosts().then((posts) => {
-      const sortedPosts = posts.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt));
-      setPosts(sortedPosts);
-      console.log('Posts:', posts);
-    });
-  }, []);
-
+const Posts = ({ posts }) => {
   return (
     <>
       {posts.map((post) => {
